requirements.txt
streamlit
pypdf
faiss-cpu
sentence-transformers
transformers
torch
